Split ui into formatters and interactive
