+++ todoman/__init__.py
@@ -1,4 +1,4 @@
-from . import version
+from todoman import version
 
 __version__ = version.version
 __documentation__ = 'https://todoman.rtfd.org/en/latest/'
